add sleep function to Lua
Lua has not any built-in sleep! What a shame.
But it's good that we always can write some C code and register it under
some global name, which is done in this commit.

Also, usleep takes microseconds, and this is not quite convenient, so
we just multiply it by 1e3.